--- a/bench/generate.exs
+++ b/bench/generate.exs
@@ -1,0 +1,6 @@
+Benchee.run(
+  %{
+    "generate" => fn -> Nanoid.NonSecure.generate() end
+  },
+  profile_after: true
+)